requirements.txt
beautifulsoup4==4.12.3
Django==4.2.1
djangorestframework==3.14.0
langchain==0.1.16
langchain_community==0.0.33
Requests==2.31.0
psycopg2==2.9.9
pillow==10.2.0
openai==1.17.0
tiktoken==0.6.0
plotly==5.21.0
sqlalchemy==2.0.28
wordcloud==1.9.3
textblob==0.18.0.post0
pandas==2.2.2

